lesson 24

+++ app/Http/Controllers/ThreadsController.php
@@ -67,9 +67,11 @@ class ThreadsController extends Controller
 
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
         $thread->delete();
 
-        if(\request()->wantsJson()) {
+        if (\request()->wantsJson()) {
             return \response([], 204);
         }
